+++ apps/festival/festival/src/app/marketplace/notification/notification.component.html
@@ -5,8 +5,8 @@
 
 <ng-template #noFeed>
   <section fxLayout="column" fxLayoutAlign="center center">
-    <h1>You have no notifications yet</h1>
-    <p>For the moment, you haven’t received any notifications.</p>
+    <h1>You have no notification yet.</h1>
+    <p>You haven't received any notifications at this time.</p>
     <img asset="no_titles.webp" alt="No title image">
   </section>
 </ng-template>
