+++ apps/festival/festival/src/app/marketplace/notification/notification.component.ts
@@ -1,6 +1,6 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { NotificationQuery } from '@blockframes/notification/+state/notification.query';
 import { map } from 'rxjs/operators';
+import { NotificationQuery } from '@blockframes/notification/+state/notification.query';
 
 
 @Component({
