+++ apps/festival/festival/src/app/marketplace/notification/notification.module.ts
@@ -3,8 +3,8 @@ import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { ImgAssetModule } from '@blockframes/ui/theme';
-import { NotificationComponent } from './notification.component';
 import { NotificationListModule } from '@blockframes/notification/components/list/list.module';
+import { NotificationComponent } from './notification.component';
 
 @NgModule({
   declarations: [NotificationComponent],
